deploy_k8_knative/requirements.txt
torch==2.1.0
torchvision==0.16.0
flask==3.0.0
pillow==10.0.1
numpy==1.24.3
gunicorn==21.2.0